[DEV-5738] moving to label to its own file
